+++ halo2_proofs/src/plonk/evaluation.rs
@@ -5,7 +5,10 @@ use std::{
 
 use ff::Field;
 use maybe_rayon::prelude::*;
-use pasta_curves::{deferred::DeferredField, pallas, vesta};
+use pasta_curves::{
+    deferred::{DeferredAccumulator, DeferredField},
+    pallas, vesta,
+};
 
 use crate::{
     arithmetic::eval_polynomial,
@@ -228,7 +231,7 @@ fn deferred_inner_product<F: DeferredField>(polynomial: &[F], powers: &[F]) -> F
     let Some((coefficient, power)) = products.next() else {
         return F::ZERO;
     };
-    let mut accumulator = F::mul_accumulator(coefficient, power);
+    let mut accumulator = F::Accumulator::initialize(coefficient, power);
     for (coefficient, power) in products {
         F::mul_accumulate(&mut accumulator, coefficient, power);
     }